Revert "Merge branch 'feature/skill-draw' into feature/player"
This reverts commit 9c24ca2faa653bef238f35231efc8c01ae8f2e7d, reversing
changes made to e3e45f97168fbe86b117e5e8f42b6aa5f82ce03b.

+++ Assets/02_Scripts/Player/PlayerModel.cs
@@ -40,7 +40,7 @@ public class PlayerModel : CharacterModelBase
     /// <summary>
     /// 플레이어 테스트용 공격력
     /// </summary>
-    public float playerAttack;
+    public int playerAttack;
     public Transform currentTarget;
 
     public void Update()
